genbase/server/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "server"
version = "0.1.0"
description = "Genbase Server"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "pydantic>=2.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "python-hcl2>=7.2.0",
]

[tool.setuptools]
packages = ["src"]




[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"

[tool.ruff]
line-length = 88
target-version = "py312"
